adding v3.0.0

+++ src/main/dcf.c
@@ -1,6 +1,6 @@
 /*
  *  R : A Computer Language for Statistical Data Analysis
- *  Copyright (C) 2001-11   The R Core Team.
+ *  Copyright (C) 2001-12   The R Core Team.
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -23,6 +23,7 @@
 
 #define R_USE_SIGNALS 1
 #include <Defn.h>
+#include <Internal.h>
 #include <Rconnections.h>
 
 #include <tre/tre.h>
@@ -127,8 +128,8 @@ SEXP attribute_hidden do_readDCF(SEXP call, SEXP op, SEXP args, SEXP env)
 			  line);
 		}
 		if(lastm >= 0) {
-		    need = strlen(CHAR(STRING_ELT(retval,
-						  lastm + nwhat * k))) + 2;
+		    need = (int) strlen(CHAR(STRING_ELT(retval,
+							lastm + nwhat * k))) + 2;
 		    if(tre_regexecb(&eblankline, line, 0, NULL, 0) == 0) {
 			is_eblankline = TRUE;
 		    } else {
@@ -142,7 +143,7 @@ SEXP attribute_hidden do_readDCF(SEXP call, SEXP op, SEXP args, SEXP env)
 			} else {
 			    offset = 0;
 			}
-			need += strlen(line + offset);
+			need += (int) strlen(line + offset);
 		    }
 		    if(buflen < need) {
 			char *tmp = (char *) realloc(buf, need);
@@ -160,7 +161,7 @@ SEXP attribute_hidden do_readDCF(SEXP call, SEXP op, SEXP args, SEXP env)
 	    } else {
 		if(tre_regexecb(&regline, line, 1, regmatch, 0) == 0) {
 		    for(m = 0; m < nwhat; m++){
-			whatlen = strlen(CHAR(STRING_ELT(what, m)));
+			whatlen = (int) strlen(CHAR(STRING_ELT(what, m)));
 			if(strlen(line) > whatlen &&
 			   line[whatlen] == ':' &&
 			   strncmp(CHAR(STRING_ELT(what, m)),
@@ -221,7 +222,7 @@ SEXP attribute_hidden do_readDCF(SEXP call, SEXP op, SEXP args, SEXP env)
 			     regmatch[0].rm_eo
 			   bytes?
 			*/
-			need = strlen(line+regmatch[0].rm_eo);
+			need = (int) strlen(line+regmatch[0].rm_eo);
 			if(buflen < need){
 			    char *tmp = (char *) realloc(buf, need);
 			    if(!tmp) {
